--- Movie List/MovieTableViewController.swift
@@ -1,55 +1,0 @@
-//
-//  MovieListTableViewController.swift
-//  Movie List
-//
-//  Created by beth on 1/24/20.
-//  Copyright © 2020 Lambda School. All rights reserved.
-//
-
-import UIKit
-
-class MovieTableViewController: UIViewController {
-
-    @IBOutlet weak var tableView: UITableView!
-
-    var movies: [Movie] = []
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        self.tableView.dataSource = self
-       
-    }
-
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "AddMovie" {
-            if let addMovieVC = segue.destination as? AddMovieViewController {
-                addMovieVC.delegate = self
-            }
-        }
-    }
-}
-
-extension MovieTableViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return movies.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: "MovieCell", for: indexPath) as? MovieTableViewCell else { fatalError() }
-
-        let movie = movies[indexPath.row]
-        cell.movie = movie
-        return cell
-    }
-}
-
-extension MovieTableViewController: AddMovieDelegate {
-    func movieWasAdded(_ movie: Movie) {
-        movies.append(movie)
-        self.navigationController?.popViewController(animated: true)
-        tableView.reloadData()
-     }
-  }
